refactor(dolphin): move dolphin manager to dolphin folder
