+++ src/main/dolphin/index.ts
@@ -1,2 +1,3 @@
 export * from "./instance";
+export * from "./manager";
 export * from "./types";
